src/backend/pyproject.toml
[project]
name = "math-solver-backend"
version = "0.1.0"
description = "Math Solver Backend using LLM"
authors = [
    {name = "Duong Dang Khoa",email = "khoadangduong63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
